Add in GestureRecognizer delegate call
This was necessary because the gesture recognizers were interfering with standard touch events that get passed to buttons.

Example:
I added a SMCalloutView to my map, which had a UIButton (Detail disclosue) on it. The TouchUpInside event would never be called because the gesture recognizers on the overlay view would interfere.

I can provide example source if necessary